chore: move missing exports
