+++ src/types/router.ts
@@ -1,4 +1,4 @@
-import { UiState } from './widget';
+import { UiState } from './ui-state';
 
 /**
  * The router is the part that saves and reads the object from the storage.
